+++ app/src/main/java/com/alongo/manyweather/presentation/geolocation_weather/GeolocationWeatherViewModel.kt
@@ -1,4 +1,4 @@
-package com.alongo.manyweather.ui.geolocation_weather
+package com.alongo.manyweather.presentation.geolocation_weather
 
 import androidx.lifecycle.ViewModel
 import com.alongo.manyweather.data.model.weather.Weather
