+++ app/src/main/res/layout/fragment_geolocation_weather.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.geolocation_weather.GeolocationWeatherFragment">
+    tools:context=".presentation.geolocation_weather.GeolocationWeatherFragment">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/buttonUpdateWeatherFragmentWeatherGeolocation"
