+++ app/src/main/res/layout/fragment_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MainFragment" >
+    tools:context=".presentation.main.MainFragment" >
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/textInputLayout"
